Title: 삼 오 무, Time: 89ms, Memory: 8MB, Status: Passed - CodeTree

+++ 231211/삼 오 무/three-five-moo.java
@@ -10,7 +10,7 @@ public class Main {
     static long n; //점의 갯수 
     
     
-    static boolean isPossible(int k) {
+    static boolean isPossible(long k) {
     	
     	long mooCount = k/3+k/5-k/15;
